Adding new file
It looks like the chapter is not just outdated but also original English version has A LOT changed than what current KR version was based on to translate. It's not simple change that sections are moved around. but I feel like this entire section is rewritten with more detailed images and explanation. I don't have time to translate it to KR right now but at least we need placeholder for the change.